requirements.txt
PyQt5
datetime
simplejson
requests
selenium
python-dateutil
psycopg2-binary
yfinance
sqlalchemy
